Create an order mode module
